Renamed core/helpers.h to util/debug.h
